third_party/blink/web_tests/svg/hixie/dynamic/001-expected.txt

ALERT: The test has passed.
A dialog should appear and say the test has passed.
